+++ amazon/containercredentials/client/src/main/kotlin/org/http4k/connect/amazon/extensions.kt
@@ -1,0 +1,7 @@
+package org.http4k.connect.amazon
+
+import org.http4k.cloudnative.env.EnvironmentKey
+import org.http4k.lens.uri
+
+
+val AWS_CONTAINER_CREDENTIALS_RELATIVE_URI = EnvironmentKey.uri().required("AWS_CONTAINER_CREDENTIALS_RELATIVE_URI")