+++ amazon/core/src/main/kotlin/org/http4k/format/AwsCoreJsonAdapterFactory.kt
@@ -2,6 +2,7 @@ package org.http4k.format
 
 import com.squareup.moshi.JsonAdapter
 import com.squareup.moshi.Moshi
+import org.http4k.connect.amazon.core.model.KotshiCredentialsJsonAdapter
 import org.http4k.connect.amazon.core.model.KotshiTagJsonAdapter
 import se.ansman.kotshi.InternalKotshiApi
 
@@ -9,7 +10,10 @@ import se.ansman.kotshi.InternalKotshiApi
 class AwsCoreJsonAdapterFactory(
     vararg typesToAdapters: Pair<String, (Moshi) -> JsonAdapter<*>>
 ) : SimpleMoshiAdapterFactory(
-    *(typesToAdapters.toList() + adapter { KotshiTagJsonAdapter() })
+    *(
+        typesToAdapters.toList()
+            + adapter { KotshiTagJsonAdapter() }
+            + adapter { moshi -> KotshiCredentialsJsonAdapter(moshi) }
+        )
         .toTypedArray()
 )
-