+++ amazon/sts/client/src/main/kotlin/org/http4k/connect/amazon/sts/action/AssumeRoleWithWebIdentity.kt
@@ -6,6 +6,8 @@ import org.http4k.connect.Http4kConnectAction
 import org.http4k.connect.RemoteFailure
 import org.http4k.connect.amazon.core.model.ARN
 import org.http4k.connect.amazon.core.model.AccessKeyId
+import org.http4k.connect.amazon.core.model.Credentials
+import org.http4k.connect.amazon.core.model.Expiration
 import org.http4k.connect.amazon.core.model.RoleSessionName
 import org.http4k.connect.amazon.core.model.SecretAccessKey
 import org.http4k.connect.amazon.core.model.SessionToken
@@ -14,9 +16,6 @@ import org.http4k.connect.amazon.core.model.WebIdentityToken
 import org.http4k.connect.amazon.core.text
 import org.http4k.connect.amazon.core.textOptional
 import org.http4k.connect.amazon.core.xmlDoc
-import org.http4k.connect.amazon.sts.model.AssumedRoleUser
-import org.http4k.connect.amazon.sts.model.Credentials
-import org.http4k.connect.amazon.sts.model.Expiration
 import org.http4k.connect.amazon.sts.model.RoleId
 import org.http4k.core.ContentType.Companion.APPLICATION_FORM_URLENCODED
 import org.http4k.core.Method.POST
@@ -83,7 +82,7 @@ data class AssumeRoleWithWebIdentity(
 }
 
 data class AssumedRoleWithWebIdentityResponse(
-    override val AssumedRoleUser: AssumedRoleUser,
+    override  val AssumedRoleId: RoleId,
     override val Credentials: Credentials,
     val SubjectFromWebIdentityToken: String,
     val Audience: String,
@@ -94,12 +93,13 @@ data class AssumedRoleWithWebIdentityResponse(
         fun from(response: Response) =
             with(response.xmlDoc()) {
                 AssumedRoleWithWebIdentityResponse(
-                    AssumedRoleUser(ARN.of(text("Arn")), RoleId.of(text("AssumedRoleId"))),
+                    RoleId.of(text("AssumedRoleId")),
                     Credentials(
                         SessionToken.of(text("SessionToken")),
                         AccessKeyId.of(text("AccessKeyId")),
                         SecretAccessKey.of(text("SecretAccessKey")),
-                        Expiration.parse(text("Expiration"))
+                        Expiration.parse(text("Expiration")),
+                        ARN.of(text("Arn"))
                     ),
                     text("SubjectFromWebIdentityToken"),
                     text("Audience"),