+++ amazon/sts/client/src/main/kotlin/org/http4k/connect/amazon/sts/action/AssumedRole.kt
@@ -1,9 +1,9 @@
 package org.http4k.connect.amazon.sts.action
 
-import org.http4k.connect.amazon.sts.model.AssumedRoleUser
-import org.http4k.connect.amazon.sts.model.Credentials
+import org.http4k.connect.amazon.core.model.Credentials
+import org.http4k.connect.amazon.sts.model.RoleId
 
 interface AssumedRole {
-    val AssumedRoleUser: AssumedRoleUser
+    val AssumedRoleId: RoleId
     val Credentials: Credentials
 }