+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 This list is not intended to be all-encompassing - it will document major and breaking API 
 changes with their rationale when appropriate. Given version `A.B.C.D`, breaking changes are to be expected in version number increments where changes in the `A` or `B` sections:
 
+### v3.17.0.0
+- **http4k-connect-*** - Upgrade dependencies.
+- **http4k-connect-amazon-core** -Added loading of credentials from Container Credentials Service.
+- **http4k-connect-amazon-sts-*** - [Unlikely break] Repackaged Credentials to core.
+
 ### v3.16.6.0
 - **http4k-connect-*** - Upgrade dependencies.